requirements.txt
openai-whisper==20250625
paddleocr==3.1.1
paddlepaddle==3.1.0
gTTS==2.5.4
python-dotenv==1.2.1